+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_reject_redelivered_messages_disabled.php
@@ -1,0 +1,7 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'messenger' => [
+        'reject_redelivered_messages' => false,
+    ],
+]);