+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_reject_redelivered_messages_disabled_explicit_bus.php
@@ -1,0 +1,14 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'messenger' => [
+        'reject_redelivered_messages' => false,
+        'default_bus' => 'messenger.bus.default',
+        'buses' => [
+            'messenger.bus.default' => null,
+            'messenger.bus.commands' => [
+                'middleware' => ['reject_redelivered_message_middleware'],
+            ],
+        ],
+    ],
+]);